requirements.txt
fastapi
uvicorn[standard]
requests
pandas
scikit-learn
joblib
python-dotenv
overpass
aiofiles
jinja2
